requirements.txt
pyTelegramBotAPI==4.10.0
requests==2.28.2
python-telegram-bot-calendar==1.0.5
python-dotenv==1.0.0
peewee==3.16.0
loguru==0.6.0
